requirements.txt
# LangChain Core
langchain
langchain-core


langchain_community


# Hugging Face Integration
langchain-huggingface


# Environment Variable Management
python-dotenv



langchain-groq

gunicorn
whitenoise
django
django-cors-headers   